Fix ImageMaker
Fix compile path in makefile
Fix type to off_t in ImageMaker

+++ OS64/04.Utility/00.ImageMaker/ImageMaker.c
@@ -107,7 +107,7 @@ int AdjustInSectorSize(int iFd, int iSourceSize){
 
 	if(iAdjustSizeToSector != 0){
 		iAdjustSizeToSector = 512 - iAdjustSizeToSector;
-		printf("[INFO] File size [%lu] and fill [%u] byte\n", iSourceSize, iAdjustSizeToSector);
+		printf("[INFO] File size [%u] and fill [%u] byte\n", iSourceSize, iAdjustSizeToSector);
 		for(i=0;i<iAdjustSizeToSector;i++){
 			write(iFd, &cCh, 1);
 		}
@@ -125,9 +125,9 @@ void WriteKernelInformation(int iTargetFd, int iTotalKernelSectorCount, int iKer
 	long lPosition;
 
 	// 5th byte of start of file = TOTALSECTORCOUNT
-	lPosition = lseek(iTargetFd, 5, SEEK_SET);
+	lPosition = lseek(iTargetFd, (off_t)5, SEEK_SET);
 	if(lPosition == -1){
-		fprintf(stderr, "lseek fail. Return value = %d, errno=%d, %d\n", lPosition, errno, SEEK_SET);
+		fprintf(stderr, "lseek fail. Return value = %ld, errno=%d, %d\n", lPosition, errno, SEEK_SET);
 		exit(-1);
 	}
 	usData = (unsigned short) iTotalKernelSectorCount;
